:sparkles: [BE] 볼트 이미지 삭제 API

See merge request s08-final/S08P31A301!79

Former-commit-id: ba14dac115e35292d74a67177bb8d6fbe1f22667

+++ backend/semes/src/main/java/com/ssafy/semes/transition/model/TransitionDeleteRequestDto.java
@@ -6,5 +6,5 @@ import lombok.Setter;
 @Setter
 @Getter
 public class TransitionDeleteRequestDto {
-	private int[] fileIds;
+	private Long[] fileIds;
 }
